Improve backend/src/repositories structure
